Added normal surface shading for sphere
Modified hitSphere() and rayColor() to allow calculation of normals and normal surface shading.
Minor changes to project folder structure, and replaced std::sqrt() calls with std::sqrtf().

+++ RayTracingIOW/src/Vec3.cpp
@@ -97,7 +97,7 @@ void Vec3::operator/=(float scalar)
 
 float Vec3::length() const
 {
-	return std::sqrt(x * x + y * y + z * z);
+	return std::sqrtf(x * x + y * y + z * z);
 }
 
 float Vec3::dot(const Vec3& other) const
